refactor: move some functionality of pro-football-reference scraper into a utility module
